Enable stateless XNNPACK linear.
The optimal solution to use XNNPACK is to separate operator creation
from execution - also called prepacking the weights.  If we have done
our job properly, JIT must have caught and replaced nn.Linear on mobile
with the prepacked versions.  Still, if we somehow end up in
at::native::linear for whatever reason, it is still more efficient to go
through XNNPACK than the alternatives of at::addmm or at::matmul.

ghstack-source-id: a2134e46101dc811b3f9793392b82e0195726b84
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35791